+++ src/novxlib/ods/ods_w_sectionlist.py
@@ -271,7 +271,7 @@ $TimeCell
 
 '''
 
-    def _get_sectionMapping(self, scId, sectionNumber, wordsTotal):
+    def _get_sectionMapping(self, scId, sectionNumber, wordsTotal, **kwargs):
         """Return a mapping dictionary for a section section.
         
         Positional arguments: