+++ src/novxlib/odt/odt_w_proof.py
@@ -29,10 +29,25 @@ class OdtWProof(OdtWFormatted):
 '''
 
     _sectionTemplate = f'''<text:h text:style-name="Heading_20_3" text:outline-level="3">$Title</text:h>
-<text:p text:style-name="section_20_mark">[$ID]</text:p>
+<text:p text:style-name="{_('Section_20_mark')}">[$ID]</text:p>
 $SectionContent
-<text:p text:style-name="section_20_mark">[/{SECTION_PREFIX}]</text:p>
+<text:p text:style-name="{_('Section_20_mark')}">[/{SECTION_PREFIX}]</text:p>
 '''
 
     _fileFooter = OdtWFormatted._CONTENT_XML_FOOTER
 
+    def _convert_from_novx(self, text, quick=False, append=False, firstInChapter=False, xml=False):
+        """Return text without markup, converted to target format.
+        
+        Positional arguments:
+            text -- string to convert.
+        
+        Optional arguments:
+            quick: bool -- if True, apply a conversion mode for one-liners without formatting.
+            append: bool -- if True, indent the first paragraph.
+            firstInChapter: bool: -- if True, the section begins a chapter.
+            xml: bool -- if True, parse XML content. 
+        
+        Overrides the superclass method.
+        """
+        return super()._convert_from_novx(text, quick=quick, append=append, firstInChapter=False, xml=xml)