+++ src/novxlib/odt/odt_writer.py
@@ -180,7 +180,7 @@ class OdtWriter(OdfFile):
  </office:settings>
 </office:document-settings>
 '''
-    _STYLES_XML = '''<?xml version="1.0" encoding="UTF-8"?>
+    _STYLES_XML = f'''<?xml version="1.0" encoding="UTF-8"?>
 
 <office:document-styles xmlns:office="urn:oasis:names:tc:opendocument:xmlns:office:1.0" xmlns:style="urn:oasis:names:tc:opendocument:xmlns:style:1.0" xmlns:text="urn:oasis:names:tc:opendocument:xmlns:text:1.0" xmlns:table="urn:oasis:names:tc:opendocument:xmlns:table:1.0" xmlns:draw="urn:oasis:names:tc:opendocument:xmlns:drawing:1.0" xmlns:fo="urn:oasis:names:tc:opendocument:xmlns:xsl-fo-compatible:1.0" xmlns:xlink="http://www.w3.org/1999/xlink" xmlns:dc="http://purl.org/dc/elements/1.1/" xmlns:meta="urn:oasis:names:tc:opendocument:xmlns:meta:1.0" xmlns:number="urn:oasis:names:tc:opendocument:xmlns:datastyle:1.0" xmlns:svg="urn:oasis:names:tc:opendocument:xmlns:svg-compatible:1.0" xmlns:chart="urn:oasis:names:tc:opendocument:xmlns:chart:1.0" xmlns:dr3d="urn:oasis:names:tc:opendocument:xmlns:dr3d:1.0" xmlns:math="http://www.w3.org/1998/Math/MathML" xmlns:form="urn:oasis:names:tc:opendocument:xmlns:form:1.0" xmlns:script="urn:oasis:names:tc:opendocument:xmlns:script:1.0" xmlns:ooo="http://openoffice.org/2004/office" xmlns:ooow="http://openoffice.org/2004/writer" xmlns:oooc="http://openoffice.org/2004/calc" xmlns:dom="http://www.w3.org/2001/xml-events" xmlns:rpt="http://openoffice.org/2005/report" xmlns:of="urn:oasis:names:tc:opendocument:xmlns:of:1.2" xmlns:xhtml="http://www.w3.org/1999/xhtml" xmlns:grddl="http://www.w3.org/2003/g/data-view#" xmlns:tableooo="http://openoffice.org/2009/table" xmlns:loext="urn:org:documentfoundation:names:experimental:office:xmlns:loext:1.0">
  <office:font-face-decls>
@@ -320,13 +320,15 @@ class OdtWriter(OdfFile):
    <style:paragraph-properties fo:margin="100%" fo:margin-left="1cm" fo:margin-right="0cm" fo:margin-top="0cm" fo:margin-bottom="0cm" fo:text-indent="0cm" style:auto-text-indent="false"/>
    <style:text-properties style:font-name="Consolas"/>
   </style:style>
-  <style:style style:name="section_20_mark" style:display-name="Section mark" style:family="paragraph" style:parent-style-name="Standard" style:next-style-name="Text_20_body" style:class="text">
+  <style:style style:name="{_('Chapter_20_beginning')}" style:display-name="{_('Chapter beginning')}" style:family="paragraph" style:parent-style-name="Text_20_body" style:next-style-name="First_20_line_20_indent" style:class="text">
+  </style:style>
+  <style:style style:name="{_('Section_20_mark')}" style:display-name="{_('Section mark')}" style:family="paragraph" style:parent-style-name="Standard" style:next-style-name="Text_20_body" style:class="text">
    <style:text-properties fo:color="#008000" fo:font-size="10pt" fo:language="zxx" fo:country="none"/>
   </style:style>
-  <style:style style:name="section_20_mark_20_unused" style:display-name="Section mark (unused type)" style:family="paragraph" style:parent-style-name="Standard" style:next-style-name="Text_20_body" style:class="text">
+  <style:style style:name="{_('Section_20_mark_20_unused')}" style:display-name="{_('Section mark unused')}" style:family="paragraph" style:parent-style-name="Standard" style:next-style-name="Text_20_body" style:class="text">
    <style:text-properties fo:color="#808080" fo:font-size="10pt" fo:language="zxx" fo:country="none"/>
   </style:style>
-  <style:style style:name="Invisible_20_Heading_20_3" style:display-name="Invisible Heading 3" style:family="paragraph" style:parent-style-name="Heading_20_3" style:class="text">
+  <style:style style:name="{_('Heading_20_3_20_invisible')}" style:display-name="{_('Heading 3 invisible')}" style:family="paragraph" style:parent-style-name="Heading_20_3" style:class="text">
    <style:paragraph-properties fo:margin-top="0cm" fo:margin-bottom="0cm" fo:line-height="100%"/>
    <style:text-properties text:display="none"/>
   </style:style>
@@ -383,7 +385,7 @@ class OdtWriter(OdfFile):
             self.novel.get_languages()
         return super().write()
 
-    def _convert_from_novx(self, text, quick=False, append=False, xml=False):
+    def _convert_from_novx(self, text, quick=False, append=False, firstInChapter=False, xml=False):
         """Return text without markup, converted to target format.
         
         Positional arguments:
@@ -392,6 +394,7 @@ class OdtWriter(OdfFile):
         Optional arguments:
             quick: bool -- if True, apply a conversion mode for one-liners without formatting.
             append: bool -- if True, indent the first paragraph.
+            firstInChapter: bool: -- if True, the section begins a chapter.
             xml: bool -- if True, parse XML content. 
         
         Overrides the superclass method.
@@ -403,7 +406,7 @@ class OdtWriter(OdfFile):
             return text
 
         if xml:
-            self._contentParser.feed(text, self.novel.languages, append)
+            self._contentParser.feed(text, self.novel.languages, append, firstInChapter)
             return ''.join(self._contentParser.odtLines)
 
         lines = text.split('\n')
@@ -427,7 +430,7 @@ class OdtWriter(OdfFile):
                 ).replace('<text:p', '\n<text:p')
         return fileHeaderMapping
 
-    def _get_sectionMapping(self, scId, sectionNumber, wordsTotal):
+    def _get_sectionMapping(self, scId, sectionNumber, wordsTotal, **kwargs):
         """Return a mapping dictionary for a section section.
         
         Positional arguments:
@@ -437,7 +440,7 @@ class OdtWriter(OdfFile):
         
         Extends the superclass method.
         """
-        sectionMapping = super()._get_sectionMapping(scId, sectionNumber, wordsTotal)
+        sectionMapping = super()._get_sectionMapping(scId, sectionNumber, wordsTotal, **kwargs)
         sectionMapping['sectionTitle'] = _('Section')
         return sectionMapping